file-generator/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
requests==2.28.1
fpdf==1.7.2
python-docx==0.8.11
openpyxl==3.0.10
flask-swagger-ui==4.11.1
flasgger==0.9.5
